Create a new runtime parameters mixin specific to model related functionality
