+++ src/distilabel/models/llms/base.py
@@ -31,7 +31,7 @@ from distilabel.errors import DistilabelNotImplementedError, DistilabelUserError
 from distilabel.exceptions import DistilabelOfflineBatchGenerationNotFinishedException
 from distilabel.mixins.runtime_parameters import (
     RuntimeParameter,
-    RuntimeParametersMixin,
+    RuntimeParametersModelMixin,
 )
 from distilabel.utils.docstring import parse_google_docstring
 from distilabel.utils.notebook import in_notebook
@@ -40,10 +40,6 @@ from distilabel.utils.serialization import _Serializable
 if TYPE_CHECKING:
     from logging import Logger
 
-    from distilabel.mixins.runtime_parameters import (
-        RuntimeParameterInfo,
-        RuntimeParametersNames,
-    )
     from distilabel.models.llms.typing import GenerateOutput, HiddenState
     from distilabel.steps.tasks.structured_outputs.outlines import StructuredOutputType
     from distilabel.steps.tasks.typing import (
@@ -59,7 +55,7 @@ if in_notebook():
     nest_asyncio.apply()
 
 
-class LLM(RuntimeParametersMixin, BaseModel, _Serializable, ABC):
+class LLM(RuntimeParametersModelMixin, BaseModel, _Serializable, ABC):
     """Base class for `LLM`s to be used in `distilabel` framework.
 
     To implement an `LLM` subclass, you need to subclass this class and implement:
@@ -240,81 +236,6 @@ class LLM(RuntimeParametersMixin, BaseModel, _Serializable, ABC):
                 raise DistilabelOfflineBatchGenerationNotFinishedException(
                     jobs_ids=self.jobs_ids  # type: ignore
                 ) from e
-
-    @property
-    def generate_parameters(self) -> List["inspect.Parameter"]:
-        """Returns the parameters of the `generate` method.
-
-        Returns:
-            A list containing the parameters of the `generate` method.
-        """
-        return list(inspect.signature(self.generate).parameters.values())
-
-    @property
-    def runtime_parameters_names(self) -> "RuntimeParametersNames":
-        """Returns the runtime parameters of the `LLM`, which are combination of the
-        attributes of the `LLM` type hinted with `RuntimeParameter` and the parameters
-        of the `generate` method that are not `input` and `num_generations`.
-
-        Returns:
-            A dictionary with the name of the runtime parameters as keys and a boolean
-            indicating if the parameter is optional or not.
-        """
-        runtime_parameters = super().runtime_parameters_names
-        runtime_parameters["generation_kwargs"] = {}
-
-        # runtime parameters from the `generate` method
-        for param in self.generate_parameters:
-            if param.name in ["input", "inputs", "num_generations"]:
-                continue
-            is_optional = param.default != inspect.Parameter.empty
-            runtime_parameters["generation_kwargs"][param.name] = is_optional
-
-        return runtime_parameters
-
-    def get_runtime_parameters_info(self) -> List["RuntimeParameterInfo"]:
-        """Gets the information of the runtime parameters of the `LLM` such as the name
-        and the description. This function is meant to include the information of the runtime
-        parameters in the serialized data of the `LLM`.
-
-        Returns:
-            A list containing the information for each runtime parameter of the `LLM`.
-        """
-        runtime_parameters_info = super().get_runtime_parameters_info()
-
-        generation_kwargs_info = next(
-            (
-                runtime_parameter_info
-                for runtime_parameter_info in runtime_parameters_info
-                if runtime_parameter_info["name"] == "generation_kwargs"
-            ),
-            None,
-        )
-
-        # If `generation_kwargs` attribute is present, we need to include the `generate`
-        # method arguments as the information for this attribute.
-        if generation_kwargs_info:
-            generate_docstring_args = self.generate_parsed_docstring["args"]
-
-            generation_kwargs_info["keys"] = []
-            for key, value in generation_kwargs_info["optional"].items():
-                info = {"name": key, "optional": value}
-                if description := generate_docstring_args.get(key):
-                    info["description"] = description
-                generation_kwargs_info["keys"].append(info)
-
-            generation_kwargs_info.pop("optional")
-
-        return runtime_parameters_info
-
-    @cached_property
-    def generate_parsed_docstring(self) -> "Docstring":
-        """Returns the parsed docstring of the `generate` method.
-
-        Returns:
-            The parsed docstring of the `generate` method.
-        """
-        return parse_google_docstring(self.generate)
 
     def get_last_hidden_states(
         self, inputs: List["StandardInput"]
